refactor(ds): update stories and components
