+++ ui/design-system-components/src/components/Profile/Bio/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import Card from '@akashaorg/design-system-core/lib/components/Card';
 import Stack from '@akashaorg/design-system-core/lib/components/Stack';
 import Text from '@akashaorg/design-system-core/lib/components/Text';
