+++ ui/design-system-components/src/components/Profile/index.ts
@@ -1,4 +1,4 @@
-export { default as ProfileLoading } from './placeholders/ProfileLoading';
+export { default as ProfileLoading } from './placeholders/profile-loading';
 export { default as ProfileHeader } from './Header';
 export { default as ProfileBio } from './Bio';
 export { default as ProfileLinks } from './Links';
