+++ ui/design-system-components/src/components/Profile/placeholders/profile-loading.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 
-import ProfileLoading from './ProfileLoading';
+import ProfileLoading from './profile-loading';
 
 const meta: Meta = {
   title: 'Profile/ProfileProfileLoading',
