+++ ui/design-system-components/src/components/SearchBar/__tests__/index.test.tsx
@@ -3,7 +3,8 @@ import { act, cleanup } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { customRender } from '@akashaorg/design-system-core/src/test-utils';
-import SearchBar from '../index';
+
+import SearchBar from '../';
 
 describe('<SearchBar /> Component', () => {
   let componentWrapper = customRender(<></>, {});
