+++ ui/design-system-components/src/components/SearchBar/search-bar.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
+
 import SearchBar, { SearchBarProps } from '.';
 
 const meta: Meta<SearchBarProps> = {
