+++ ui/design-system-components/src/components/SwitchCard/__tests__/index.test.tsx
@@ -2,8 +2,9 @@ import * as React from 'react';
 import { act, cleanup } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import SwitchCard from '..';
 import { customRender } from '@akashaorg/design-system-core/src/test-utils';
+
+import SwitchCard from '..';
 
 describe('<SwitchCard /> Component', () => {
   let componentWrapper = customRender(<></>, {});
